Chemical concentration rounded to nearest int.
Reduce the steepness of the gradient.

+++ Assets/Scripts/Chemical.cs
@@ -13,9 +13,9 @@ public class Chemical : MonoBehaviour {
         origin = transform.position;
     }
 
-    public float getConcentration(Vector3 position)
+    public int getConcentration(Vector3 position)
     {
-        return transform.localScale.magnitude / Vector3.Distance(origin, position);
+        return (int) (transform.localScale.magnitude / Vector3.Distance(origin, position));
     }
 
     public void setChemotaxisType(bool chemotaxisType)
